split up gpu tests
